Refactored pwHandling into dust
